+++ common/src/main/java/store/_0982/common/domain/sellerbalance/SellerBalance.java
@@ -1,4 +1,4 @@
-package store._0982.common.domain.settlement;
+package store._0982.common.domain.sellerbalance;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -22,7 +22,7 @@ import java.util.UUID;
 public class SellerBalance {
 
     @Id
-    @Column(name = "balance_id", nullable = false, updatable = false)
+    @Column(name = "balance_id", nullable = false)
     private UUID balanceId;
 
     @Column(name = "member_id", nullable = false, unique = true)
@@ -32,7 +32,7 @@ public class SellerBalance {
     private Long settlementBalance;
 
     @CreationTimestamp
-    @Column(name = "created_at", nullable = false, updatable = false)
+    @Column(name = "created_at", nullable = false)
     private OffsetDateTime createdAt;
 
     @UpdateTimestamp
