+++ common/src/main/java/store/_0982/common/domain/sellerbalance/SellerBalanceHistory.java
@@ -1,4 +1,4 @@
-package store._0982.common.domain.settlement;
+package store._0982.common.domain.sellerbalance;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
