+++ common/src/main/java/store/_0982/common/domain/settlement/SellerPayout.java
@@ -1,4 +1,4 @@
-package store._0982.common.domain.sellerpayout;
+package store._0982.common.domain.settlement;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
