+++ common/src/main/java/store/_0982/common/domain/settlement/SellerPayoutStatus.java
@@ -1,4 +1,4 @@
-package store._0982.common.domain.sellerpayout;
+package store._0982.common.domain.settlement;
 
 public enum SellerPayoutStatus {
     PENDING,
